+++ src/MaxxECU.h
@@ -14,58 +14,56 @@
 * •Requires MaxxECU Firmware version 1.67 or newer.
 */
 
-namespace Maxx
-{
-	struct Signals
-	{
-		int RPM;                                // Factor 1, unit rpm
-		int TPS;                                // Factor 0.1, unit %
-		int MAP;                                // Factor 0.1, unit kPa
-		int Lambda;                             // Factor 0.001, unit lambda
-		int LambdaA;                            // Factor 0.001, unit lambda
-		int LambdaB;                            // Factor 0.001, unit lambda
-		int IgnitionAngle;                      // Factor 0.1, unit degrees BTDC
-		int IgnitionCut;                        // Factor 1, unit %
-		int FuelPulsewidth;                     // Factor 0.01, unit ms
-		int FuelDuty;                           // Factor 0.1, unit %
-		int FuelCut;                            // Factor 1, unit %
-		int VehicleSpeed;                       // Factor 0.1, unit km/h
-		int UndrivenWheelsAvgSpeed;             // Factor 0.1, unit km/h
-		int DrivenWheelsAvgSpeed;               // Factor 0.1, unit km/h
-		int WheelSlip;                          // Factor 0.1, unit %
-		int TargetSlip;                         // Factor 0.1, unit %
-		int TractionCtrlPwrLimit;               // Factor 0.1, unit %
-		int LambdaCorrA;                        // Factor 0.1, unit %
-		int LambdaCorrB;                        // Factor 0.1, unit %
-		int FirmwareVersion;                    // Factor 0.01 
-		int BatteryVoltage;                     // Factor 0.01, unit Volt
-		int BaroPressure;                       // Factor 0.1, unit kPa
-		int IntakeAirTemp;                      // Factor 0.1, unit Celsius
-		int CoolantTemp;                        // Factor 0.1, unit Celsius
-		int TotalFuelTrim;                      // Factor 0.1, unit %
-		int EthanolConcentration;               // Factor 0.1, unit %
-		int TotalIgnitionComp;                  // Factor 0.1, unit degrees
-		int EGT1;                               // Factor 1, unit Celsius
-		int EGT2;                               // Factor 1, unit Celsius
-		int EGT3;                               // Factor 1, unit Celsius
-		int EGT4;                               // Factor 1, unit Celsius
-		int EGT5;                               // Factor 1, unit Celsius
-		int EGT6;                               // Factor 1, unit Celsius
-		int EGT7;                               // Factor 1, unit Celsius
-		int EGT8;                               // Factor 1, unit Celsius
-		int EGTHighest;                         // Factor 1, unit Celsius
-		int EGTDifference;                      // Factor 1, unit Celsius
-		int CPUTemp;                            // Factor 1, unit Celsius
-		int ErrorCodeCount;                     // Factor 1
-		int LostSyncCount;                      // Factor 1
-		int UserAnalogInput1;                   // Factor 0.1
-		int UserAnalogInput2;                   // Factor 0.1
-		int UserAnalogInput3;                   // Factor 0.1
-		int UserAnalogInput4;                   // Factor 0.1
-		int Gear;                               // Factor 1
-		int BoostSolenoidDuty;                  // Factor 0.1, unit %
-		int Spare1;                             // Factor 1
-		int Spare2;                             // Factor 1 
+namespace Maxx{
+	struct Signals{
+		int16_t RPM;                                // Factor 1, unit rpm
+		int16_t TPS;                                // Factor 0.1, unit %
+		int16_t MAP;                                // Factor 0.1, unit kPa
+		int16_t Lambda;                             // Factor 0.001, unit lambda
+		int16_t LambdaA;                            // Factor 0.001, unit lambda
+		int16_t LambdaB;                            // Factor 0.001, unit lambda
+		int16_t IgnitionAngle;                      // Factor 0.1, unit degrees BTDC
+		int16_t IgnitionCut;                        // Factor 1, unit %
+		int16_t FuelPulsewidth;                     // Factor 0.01, unit ms
+		int16_t FuelDuty;                           // Factor 0.1, unit %
+		int16_t FuelCut;                            // Factor 1, unit %
+		int16_t VehicleSpeed;                       // Factor 0.1, unit km/h
+		int16_t UndrivenWheelsAvgSpeed;             // Factor 0.1, unit km/h
+		int16_t DrivenWheelsAvgSpeed;               // Factor 0.1, unit km/h
+		int16_t WheelSlip;                          // Factor 0.1, unit %
+		int16_t TargetSlip;                         // Factor 0.1, unit %
+		int16_t TractionCtrlPwrLimit;               // Factor 0.1, unit %
+		int16_t LambdaCorrA;                        // Factor 0.1, unit %
+		int16_t LambdaCorrB;                        // Factor 0.1, unit %
+		int16_t FirmwareVersion;                    // Factor 0.01 
+		int16_t BatteryVoltage;                     // Factor 0.01, unit Volt
+		int16_t BaroPressure;                       // Factor 0.1, unit kPa
+		int16_t IntakeAirTemp;                      // Factor 0.1, unit Celsius
+		int16_t CoolantTemp;                        // Factor 0.1, unit Celsius
+		int16_t TotalFuelTrim;                      // Factor 0.1, unit %
+		int16_t EthanolConcentration;               // Factor 0.1, unit %
+		int16_t TotalIgnitionComp;                  // Factor 0.1, unit degrees
+		int16_t EGT1;                               // Factor 1, unit Celsius
+		int16_t EGT2;                               // Factor 1, unit Celsius
+		int16_t EGT3;                               // Factor 1, unit Celsius
+		int16_t EGT4;                               // Factor 1, unit Celsius
+		int16_t EGT5;                               // Factor 1, unit Celsius
+		int16_t EGT6;                               // Factor 1, unit Celsius
+		int16_t EGT7;                               // Factor 1, unit Celsius
+		int16_t EGT8;                               // Factor 1, unit Celsius
+		int16_t EGTHighest;                         // Factor 1, unit Celsius
+		int16_t EGTDifference;                      // Factor 1, unit Celsius
+		int16_t CPUTemp;                            // Factor 1, unit Celsius
+		int16_t ErrorCodeCount;                     // Factor 1
+		int16_t LostSyncCount;                      // Factor 1
+		int16_t UserAnalogInput1;                   // Factor 0.1
+		int16_t UserAnalogInput2;                   // Factor 0.1
+		int16_t UserAnalogInput3;                   // Factor 0.1
+		int16_t UserAnalogInput4;                   // Factor 0.1
+		int16_t Gear;                               // Factor 1
+		int16_t BoostSolenoidDuty;                  // Factor 0.1, unit %
+		int16_t Spare1;                             // Factor 1
+		int16_t Spare2;                             // Factor 1 
 	};
 
 	void initStruct(Signals& mySignals);
